LR- moved category reducers and actions in category folder and changed necessary paths.
