+++ resources/assets/js/actions/authentication/user.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
-import { APP_ENDPOINT } from '../../constants';
-import { USER_FETCHED, USER_FETCHED_FAILURE } from '../../action-types';
-import { loggedInUser } from '../../services/urls/index.js';
+import { APP_ENDPOINT } from 'js/constants';
+import { USER_FETCHED, USER_FETCHED_FAILURE } from 'js/action-types';
+import { loggedInUser } from 'js/services/urls/index.js';
 
 const userFetched = user => ({
     type: USER_FETCHED,
