+++ resources/assets/js/actions/category/category-modal.js
@@ -1,7 +1,7 @@
 import {
     CATEGORY_ADD_MODAL_OPENED,
     CATEGORY_ADD_MODAL_RESET
-} from '../action-types';
+} from 'js/action-types';
 
 export const opened = () => ({ type: CATEGORY_ADD_MODAL_OPENED });
