+++ resources/assets/js/actions/category/clear-category-state.js
@@ -1,4 +1,4 @@
-import { CATEGORY_STATE_CLEARED } from '../action-types';
+import { CATEGORY_STATE_CLEARED } from 'js/action-types';
 
 export const clearCategoryState = () => ({
     type: CATEGORY_STATE_CLEARED
