--- resources/assets/js/components/test.js
@@ -1,19 +1,0 @@
-import React from 'react';
-
-class Test extends React.Component {
-	constructor(props) {
-		super(props);
-		console.log('123');
-	}
-	render() {
-		return (
-			<div id="wrapper">
-		        <h1>Test Swapnil</h1>
-		        <a href="categories"> Test</a>
-		    </div>
-		);
-	}
-}
-    
-
-export default Test;
