+++ resources/assets/js/reducers/index.js
@@ -1,8 +1,8 @@
-export { default as loginEditor } from './authetication/login-editor';
-export { default as auth } from './authetication/auth';
-export { default as user } from './authetication/user';
-export { default as categories } from './categories';
-export { default as categoryEditor } from './category-editor';
-export { default as category } from './category';
-export { default as modal } from './modal';
-export { default as loadingState } from './loading-state';
+export { default as loginEditor } from 'js/reducers/authetication/login-editor';
+export { default as auth } from 'js/reducers/authetication/auth';
+export { default as user } from 'js/reducers/authetication/user';
+export { default as categories } from 'js/reducers/category/categories';
+export { default as categoryEditor } from 'js/reducers/category/category-editor';
+export { default as category } from 'js/reducers/category/category';
+export { default as modal } from 'js/reducers/category/modal';
+export { default as loadingState } from 'js/reducers/loading-state';
