+++ resources/assets/js/reducers/loading-state.js
@@ -1,4 +1,4 @@
-import { LOADING_STATE_ENABLED, LOADING_STATE_DISABLED } from '../action-types';
+import { LOADING_STATE_ENABLED, LOADING_STATE_DISABLED } from 'js/action-types';
 
 const initialState = false;
